+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Zephyr] Added a tool `create-test-case` for creating Test Case [#320](https://github.com/SmartBear/smartbear-mcp/pull/320)
 - [Zephyr] Added a tool `create-test-cycle` for creating Test Cycle [#323](https://github.com/SmartBear/smartbear-mcp/pull/323)
+- [Zephyr] Added a tool `create-test-script` for creating Test Script [#328](https://github.com/SmartBear/smartbear-mcp/pull/328)
 
 ## [0.13.5] - 2026-02-02